test: share portable renderer coverage
