+++ src/lib.rs
@@ -38,9 +38,12 @@ mod gitignore_tests;
 #[cfg(test)]
 #[path = "../tests/crate/icons.rs"]
 mod icons_tests;
-#[cfg(all(test, unix))]
+#[cfg(test)]
 #[path = "../tests/crate/render.rs"]
 mod render_tests;
+#[cfg(all(test, unix))]
+#[path = "../tests/crate/render_unix.rs"]
+mod render_unix_tests;
 #[cfg(all(test, windows))]
 #[path = "../tests/crate/render_windows.rs"]
 mod render_windows_tests;
