+++ tests/crate/common.rs
@@ -1,7 +1,6 @@
 use colored_text::{ColorMode, ColorizeConfig};
 use std::sync::{Mutex, MutexGuard};
 
-#[cfg(unix)]
 use crate::Params;
 
 static COLOR_MODE_LOCK: Mutex<()> = Mutex::new(());
@@ -31,7 +30,6 @@ impl Drop for ColorModeGuard {
     }
 }
 
-#[cfg(unix)]
 pub(crate) fn has_ansi(text: &str) -> bool {
     text.contains("\u{1b}[")
 }
@@ -43,7 +41,6 @@ pub(crate) fn with_color_output_enabled<T>(test: impl FnOnce() -> T) -> T {
     })
 }
 
-#[cfg(unix)]
 pub(crate) fn fixed_time_params() -> Params {
     Params {
         time_gradient: false,
@@ -51,7 +48,6 @@ pub(crate) fn fixed_time_params() -> Params {
     }
 }
 
-#[cfg(unix)]
 pub(crate) fn plain_permission_params() -> Params {
     Params {
         permission_colors: false,
@@ -60,7 +56,6 @@ pub(crate) fn plain_permission_params() -> Params {
     }
 }
 
-#[cfg(unix)]
 pub(crate) fn accentless_params() -> Params {
     Params {
         permission_colors: false,
@@ -70,7 +65,6 @@ pub(crate) fn accentless_params() -> Params {
     }
 }
 
-#[cfg(unix)]
 pub(crate) fn time_only_params() -> Params {
     Params {
         permission_colors: false,
